+++ api/controllers/gfe_typing_controller.py
@@ -1,40 +1,40 @@
+from pygfe.models.error import Error
 from pygfe.pygfe import pyGFE
+
+from seqann.sequence_annotation import BioSeqAnn
+
 from py2neo import Graph
 from neo4j.exceptions import ServiceUnavailable
-from pygfe.models.error import Error
+
+from pandas import DataFrame
+
 import logging
 import io
 import re
 import yaml
-from pandas import DataFrame
-from seqann.sequence_annotation import BioSeqAnn
-from pygfe.models.feature import Feature
 
 seqanns = {}
 gfe_feats = None
 gfe2hla = None
 seq2hla = None
-neo_file = open("swagger_server/neo4j.yaml", "r")
+neo_file = open("neo4j.yaml", "r")
 neo_dict = yaml.safe_load(neo_file)
 
 
-def gfecreate_post(locus, sequence, imgt_version, neo4j_url=neo_dict['neo4j_url'], user=neo_dict['user'],
-                   password=neo_dict['password']):  # noqa: E501
-    """gfecreate_post
+def gfeTyping_get(sequence, locus, imgthla_version="3.31.0", neo4j_url=neo_dict['neo4j_url'], user=neo_dict['user'], password=neo_dict['password']):  # noqa: E501
+    """gfeTyping_get
 
-    Get all features associated with a locus
+    Get HLA and GFE from consensus sequence or GFE notation
 
     :param locus: Valid HLA locus
-    :param sequence: Valid sequence
-    :param imgt_version : db version
+    :param sequence: Consensus sequence
+    :param imgthla_version: IMGT/HLA DB Version
     :rtype: Typing
     """
-    imgthla_version = imgt_version
     global seqanns
     global gfe_feats
     global gfe2hla
     global seq2hla
-    pygfe = pyGFE()
     sequence = sequence['sequence']
     log_capture_string = io.StringIO()
     logger = logging.getLogger('')
@@ -43,8 +43,9 @@ def gfecreate_post(locus, sequence, imgt_version, neo4j_url=neo_dict['neo4j_url'
 
     # create console handler and set level to debug
     ch = logging.StreamHandler(log_capture_string)
-    formatter = logging.Formatter('%(asctime)s - %(name)-35s - %(levelname)-5s'
-                                  ' - %(funcName)s %(lineno)d: - %(message)s')
+    formatter = logging.Formatter(
+        '%(asctime)s - %(name)-35s - %(levelname)-5s '
+        '- %(funcName)s %(lineno)d: - %(message)s')
     ch.setFormatter(formatter)
     ch.setLevel(logging.INFO)
     logger.addHandler(ch)
@@ -61,6 +62,7 @@ def gfecreate_post(locus, sequence, imgt_version, neo4j_url=neo_dict['neo4j_url'
         seqann = BioSeqAnn(verbose=True, safemode=True,
                            dbversion=db, verbosity=3)
         seqanns.update({db: seqann})
+
     try:
         graph = Graph(neo4j_url,
                       user=user,
@@ -86,9 +88,9 @@ def gfecreate_post(locus, sequence, imgt_version, neo4j_url=neo_dict['neo4j_url'
                       gfe_feats=gfe_feats,
                       seq2hla=seq2hla,
                       verbose=True)
+
     try:
-        typing = pygfe.gfe_create(locus=locus, sequence=sequence,
-                                  imgtdb_version=db)
+        typing = pygfe.type_from_seq(locus, sequence, imgthla_version)
     except Exception as e:
         print(e)
         log_contents = log_capture_string.getvalue()
@@ -102,20 +104,7 @@ def gfecreate_post(locus, sequence, imgt_version, neo4j_url=neo_dict['neo4j_url'
 
     if not typing:
         log_contents = log_capture_string.getvalue()
-        return Error("Type with alignment failed",
-                     log=log_contents.split("\n")), 404
-    structute_feats = []
-    for f in typing['structure']:
-        fn = Feature(accession=f.accession, rank=f.rank,
-                     term=f.term, sequence=f.sequence)
-        structute_feats.append(fn)
-    anno_feats = []
-    for f in typing['annotation'].structure:
-        fn = Feature(accession=f.accession, rank=f.rank,
-                     term=f.term, sequence=f.sequence)
-        anno_feats.append(fn)
-    return {
-        'gfe': typing['gfe'],
-        'feature': structute_feats,
-        'annotation_feature': anno_feats
-    }
+        return Error("Type sequence failed", log=log_contents.split("\n")), 404
+
+    typing.gfedb_version = "2.0.0"
+    return typing
