+++ api/tests/__init__.py
@@ -2,13 +2,14 @@ import logging
 
 import connexion
 from flask_testing import TestCase
-from swagger_server.encoder import JSONEncoder
+from api.encoder import JSONEncoder
+
 
 class BaseTestCase(TestCase):
 
     def create_app(self):
         logging.getLogger('connexion.operation').setLevel('ERROR')
-        app = connexion.App(__name__, specification_dir='../swagger/')
+        app = connexion.App(__name__, specification_dir='../../spec/')
         app.app.json_encoder = JSONEncoder
-        app.add_api('swagger.yaml')
+        app.add_api('gfe_service_api_spec.yaml')
         return app.app
