+++ api/tests/gfe_create_controller_test.py
@@ -1,15 +1,17 @@
-from swagger_server.tests import BaseTestCase
+from api.tests import BaseTestCase
 import json
 
 
-class TestTyping(BaseTestCase):
+class TestGfeCreation(BaseTestCase):
 
-    def test_typing(self):
-        """Test case for typing
+    def test_gfe_creation(self):
+        """Test case for gfe_creation
         """
         locus = 'HLA-DQB1'
-        response = self.client.post(
-            f'/gfe/typing/hla/{locus}',
+        imgt_version = '3.31.0'
+        response = self.client.open(
+            f'/gfe/{locus}/{imgt_version}',
+            method='PUT',
             data=json.dumps(dict(sequence="AGAACGGGAAGGAGACGCTGCAGCGCACGGGTACCAGGGGCCACGGGGCGCCTCCCTGAT")),
             content_type='application/json'
         )
