+++ api/tests/gfe_notation_controller_test.py
@@ -1,4 +1,4 @@
-from swagger_server.tests import BaseTestCase
+from api.tests import BaseTestCase
 import json
