+++ main.py
@@ -2,13 +2,12 @@
 
 import connexion
 from flask import redirect
-from swagger_server import encoder
+from api import encoder
 
-app = connexion.App(__name__, specification_dir='./swagger_server/swagger/')
+app = connexion.App(__name__, specification_dir='spec/')
 app.app.json_encoder = encoder.JSONEncoder
-app.add_api('swagger.yaml', arguments={'title': 'gfe rest services'})
+app.add_api('gfe_service_api_spec.yaml', arguments={'title': 'GFE Services'})
 app.app.config['MAX_CONTENT_LENGTH'] = 1600 * 1024 * 1024
-
 
 
 @app.route("/")
