+++ requirements.txt
@@ -1,9 +1,9 @@
-connexion[swagger-ui]==2.7.0
 Flask==1.1.2
 Flask-Testing==0.8.1
 py-gfe==1.1.2
 pytest==6.2.2
 seq-ann==1.1.0
+connexion[swagger-ui]==2.7.0
 swagger-spec-validator==2.7.3
 uWSGI==2.0.19.1
 Werkzeug==1.0.1
